requirements.txt
flask
requests
stomp.py